% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Função para dividir}
\usage{
dividir(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Um número}
}
\description{
Essa função serve para dividir dois números
}
\examples{
dividir(a, b)

}
