% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{divisao}
\alias{divisao}
\title{Função para dividir dois números}
\usage{
divisao(a, b)
}
\arguments{
\item{a}{o número do qual será dividido (dividendo)}

\item{b}{o número que informa a quantidade de partes (divisor)}
}
\value{
A divisão entre \code{a} e \code{b}
}
\description{
Esta função recebe dois valores numéricos e devolve a divisão do primeiro pelo segundo
}
\examples{
 divisao(8,4)
 divisao(10,2)
}
