% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historico_operacoes}
\alias{historico_operacoes}
\title{Histórico de Operações da Calculadora}
\format{
Um data frame com 4 linhas e 5 variáveis:
\describe{
  \item{id}{Identificador único da operação}
  \item{operacao}{O nome da operação realizada (soma, subtracao, multiplicacao, divisao)}
  \item{valor1}{O primeiro valor numérico da conta}
  \item{valor2}{O segundo valor numérico da conta}
  \item{resultado}{O resultado final da operação matemática}
}
}
\source{
Dados simulados para fins de aprendizagem no RStudio.
}
\usage{
historico_operacoes
}
\description{
Um banco de dados simulado contendo um histórico de operações matemáticas
básicas para demonstração do pacote.
}
\keyword{datasets}
