% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{multiplicacao}
\alias{multiplicacao}
\title{Função para multiplicar dois números}
\usage{
multiplicacao(a, b)
}
\arguments{
\item{a}{o número do qual será multiplicado}

\item{b}{o número do qual será multiplicado}
}
\value{
A multiplicação entre \code{a} e \code{b}
}
\description{
Esta função recebe dois valores numéricos e devolve a multipicação do primeiro pelo segundo
}
\examples{
 multiplicacao(10,4)
 multiplicacao(5,6)
}
