% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Função para multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Um número}
}
\description{
Essa função serve para multiplicar dois números
}
\examples{
multiplicar(1, 1)


}
