% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{soma}
\alias{soma}
\title{Função para somar dois números}
\usage{
soma(a, b)
}
\arguments{
\item{a}{Um número.}

\item{b}{Outro número.}
}
\value{
A soma de \code{a} e \code{b}.
}
\description{
Esta função recebe dois valores numéricos e devolve a soma de ambos.
}
\examples{
soma(2, 3)
soma(10, -5)
}
