% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Somar}
\usage{
somar(x, y)
}
\arguments{
\item{x}{um valor numerico}

\item{y}{um valor numerico}
}
\value{
x + y
}
\description{
Esta funcao retorna a soma de dois numeros.
}
\examples{
somar(1, 1)

}
