% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacoes.R
\name{subtracao}
\alias{subtracao}
\title{Função para subtrair dois números}
\usage{
subtracao(a, b)
}
\arguments{
\item{a}{O número do qual será subtraído (minuendo).}

\item{b}{O valor a ser subtraído (subtraendo).}
}
\value{
A diferença entre \code{a} e \code{b}.
}
\description{
Esta função recebe dois valores numéricos e devolve a subtração do primeiro pelo segundo.
}
\examples{
subtracao(10, 3)
subtracao(5, 8)
}
