% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Subtrair}
\usage{
subtrair(x, y)
}
\arguments{
\item{x}{um valor numerico}

\item{y}{um valor numerico}
}
\value{
x - y
}
\description{
Esta funcao retorna a subtracao de dois numeros.
}
\examples{
subtrair(10, 5)

}
