% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Função para Dividir}
\usage{
dividir(a, b)
}
\arguments{
\item{a}{um número}

\item{b}{outro número}
}
\description{
Esta função serve para dividir dois números
}
\examples{
dividir(5, 2)

}
