% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Função para Multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{um número}

\item{b}{outro número}
}
\description{
Esta função serve para multiplicar dois números
}
\examples{
multiplicar(5, 2)

}
