% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Um número}
}
\description{
Essa função serve para somar dois números
}
\examples{
somar(1, 1)


}
