% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Função Somar
Esta função faz a soma de dois valores}
\usage{
a um número para somar

b outro número para somar
}
\description{
Função Somar
Esta função faz a soma de dois valores
}
\examples{
somar(3,4)
}
