% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Outro número}
}
\description{
Esta função serve para somar dois números
}
\examples{
somar(3, 4)

}
