% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{número inteiro}

\item{b}{número inteiro}
}
\description{
uma função que subtrai dois números
}
\examples{
subtrair(1,1)
}
