% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R, R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para somar}
\usage{
subtrair(a, b)

subtrair(a, b)
}
\arguments{
\item{a}{Um número}

\item{b}{Outro número}
}
\description{
Esta função serve para somar dois números

Esta função serve para subtrair dois números
}
\examples{
somar(3,4)

subtrair(3,4)

}
