% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair.}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{um número.}

\item{B}{outro número}
}
\description{
Essa subtrair dois números.
}
\examples{
somar(8,3)

}
