% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função Subtrair
Esta função subtrai a soma de dois valores}
\usage{
a um número para subtrair

b outro número para subtrair
}
\description{
Função Subtrair
Esta função subtrai a soma de dois valores
}
\examples{
subtrair(3,4)
}
