% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{um número}

\item{b}{outro número}
}
\description{
Esta função serve para subtrair dois números
}
\examples{
subtrair(3, 4)

}
